MSF-89 - Refactoring server factory flow.
