+++ Factory/src/main/kotlin/net/milosvasic/factory/mail/application/server_factory/ServerFactoryInitializationOperation.kt
@@ -1,0 +1,5 @@
+package net.milosvasic.factory.mail.application.server_factory
+
+import net.milosvasic.factory.mail.common.initialization.InitializationOperation
+
+class ServerFactoryInitializationOperation : InitializationOperation()
