+++ Factory/src/main/kotlin/net/milosvasic/factory/mail/common/initialization/TerminationOperation.kt
@@ -1,0 +1,5 @@
+package net.milosvasic.factory.mail.common.initialization
+
+import net.milosvasic.factory.mail.operation.Operation
+
+abstract class TerminationOperation : Operation()
