+++ Factory/src/main/kotlin/net/milosvasic/factory/mail/component/docker/DockerInitializationOperation.kt
@@ -1,5 +1,6 @@
 package net.milosvasic.factory.mail.component.docker
 
+import net.milosvasic.factory.mail.common.initialization.InitializationOperation
 import net.milosvasic.factory.mail.operation.Operation
 
-class DockerInitializationOperation : Operation()
+class DockerInitializationOperation : InitializationOperation()
