+++ Factory/src/main/kotlin/net/milosvasic/factory/mail/operation/OperationResult.kt
@@ -2,9 +2,14 @@ package net.milosvasic.factory.mail.operation
 
 import net.milosvasic.factory.mail.EMPTY
 
-data class OperationResult(
+open class OperationResult(
     val operation: Operation,
     val success: Boolean,
     val data: String = String.EMPTY,
     val exception: Exception? = null
-)
+) {
+
+    override fun toString(): String {
+        return "OperationResult(operation=$operation, success=$success, data='$data', exception=$exception)"
+    }
+}
